split stdlib to any files
